+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## [0.2.7] - Unreleased
-
+- Added guard against custom subtype in random ISCC-CODE generation.
 
 ## [0.2.6] - 2022-04-13
 - Added `KY` and `MM` to valid prefixes